fix(turn): split world-market pipeline fixtures under 400-line gate (#4039)
Extract minor/purchased timber pipeline Game builders so world_market_test_support.dart stays within repo.domain_package_test_file_size.
