+++ packages/colonizethis_turn/test/support/world_market_pipeline_game_fixtures.dart
@@ -1,0 +1,143 @@
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import 'riches_to_treasury_phase_purchased_tile_riches_test_support.dart';
+
+Game minorTimberAutoOfferPipelineGame({
+  required int buyerTreasury,
+  int timberPrice = 30,
+}) {
+  const ow = 'oldWorld';
+  const minorProvinceId = '$ow|m1';
+  const tileKey = '$ow|m1|0|0';
+  return Game(
+    id: 'g_c7_minor_timber',
+    players: [
+      Player(
+        id: 'gpBuyer',
+        displayName: 'Buyer',
+        isHuman: false,
+        stockpile: Stockpile.empty,
+        treasury: buyerTreasury,
+      ),
+    ],
+    minorNations: const [
+      MinorNation(
+        id: 'm1',
+        capitalProvinceId: minorProvinceId,
+        capitalTile: CapitalTile(
+          regionId: ow,
+          provinceId: minorProvinceId,
+          x: 0,
+          y: 0,
+        ),
+      ),
+    ],
+    worldState: WorldState(
+      turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
+      oldWorld: RegionData(
+        provinces: [
+          Province(
+            id: minorProvinceId,
+            regionId: ow,
+            ownerId: 'm1',
+            townDevelopmentLevel: 1,
+          ),
+        ],
+      ),
+      newWorld: const RegionData(),
+      tileKeysByRegionAndProvince: const {
+        ow: {
+          minorProvinceId: [tileKey],
+        },
+      },
+      tileState: TileMapState()
+          .setImprovement(tileKey, 1)
+          .setRoadLevel(tileKey, 1),
+    ),
+    worldMarketState: WorldMarketState.empty.copyWith(
+      prices: {'timber': timberPrice},
+    ),
+  );
+}
+
+Map<String, TileMapResult> minorTimberTileMapByRegion() =>
+    tileMapByRegionForResource(Resource.timber);
+
+Game purchasedTimberBidPipelineGame({
+  required int gpATreasury,
+  int timberPrice = 25,
+}) {
+  const ow = 'oldWorld';
+  const minorProvinceId = '$ow|M1';
+  const tileKey = '$ow|M1|0|0';
+  return Game(
+    id: 'g_c7_purchased_timber',
+    players: [
+      Player(
+        id: 'gpA',
+        displayName: 'GP A',
+        isHuman: true,
+        treasury: gpATreasury,
+        stockpile: Stockpile.empty,
+      ),
+    ],
+    minorNations: const [
+      MinorNation(
+        id: 'M1',
+        displayName: 'Minor 1',
+        capitalProvinceId: minorProvinceId,
+        capitalTile: CapitalTile(
+          regionId: ow,
+          provinceId: minorProvinceId,
+          x: 0,
+          y: 0,
+        ),
+      ),
+    ],
+    worldState: WorldState(
+      turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
+      oldWorld: RegionData(
+        provinces: [
+          Province(
+            id: minorProvinceId,
+            regionId: ow,
+            ownerId: 'M1',
+            townDevelopmentLevel: 1,
+          ),
+        ],
+      ),
+      newWorld: const RegionData(),
+      tileKeysByRegionAndProvince: const {
+        ow: {
+          minorProvinceId: [tileKey],
+        },
+      },
+      purchasedTilesByTileKey: const {tileKey: 'gpA'},
+      tileState: TileMapState()
+          .setImprovement(tileKey, 1)
+          .setRoadLevel(tileKey, 1),
+    ),
+    worldMarketState: WorldMarketState.empty.copyWith(
+      prices: {'timber': timberPrice},
+    ),
+  );
+}
+
+Orders timberBidOrdersForGp({
+  required String gpId,
+  int quantity = 1,
+  int priority = 1,
+}) =>
+    Orders(
+      tradeOrdersByPlayerId: {
+        gpId: [
+          TradeOrder(
+            commodityId: 'timber',
+            type: TradeOrderType.bid,
+            quantity: quantity,
+            priority: priority,
+          ),
+        ],
+      },
+    );
